Add files via upload
A random name generator for when you have zero inspiration and feel like letting something doing the work for you. You can use it and get plenty of weird names to use for games/apps. The Minimum Syllable amount is 2, so I'm sorry for all of you who wanted 1 letter names. :c